+++ feature/main/src/main/java/com/example/main/component/CollegeCardView.kt
@@ -1,0 +1,24 @@
+package com.example.main.component
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.msg.design_system.theme.BitgoeulAndroidTheme
+import com.msg.model.ui.CollegeData
+
+@Composable
+fun CollegeCardView(
+    modifier: Modifier = Modifier,
+    data: CollegeData
+) {
+    BitgoeulAndroidTheme { colors, typography ->
+        Card(
+            modifier = modifier
+                .fillMaxWidth()
+        ) {
+
+        }
+    }
+}
